pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dedelayed"
version = "0.0.1"
description = ""
readme = "README.md"
authors = [
    { name = "Mateen Ulhaq", email = "mateen.ulhaq@interdigital.com" },
    { name = "InterDigital AI Lab", email = "compressai@interdigital.com" },
]
requires-python = ">=3.8,<3.12"

dependencies = [
    "datasets>=3.0.0",
    "efficientvit",
    "einops>=0.8.0",
    "hydra-core>=1.3.2",
    "numpy>=1.24.4",
    "omegaconf>=2.3.0",
    "opencv-python>=4.10.0.84",
    "pillow>=10.0.0",
    "setuptools>=75.3.0",
    "timm>=1.0.0",
    "torch>=1.13.1",
    "torchvision>=0.14.1",
    "transformers>=4.46.0",
    "typing-extensions>=4.12.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
secondary = [
    "av>=12.3.0",
    "fastprogress>=1.0.0",
    "hsluv>=5.0.0",
    "matplotlib>=3.7.5",
    "onnx>=1.17.0",
    "torchmetrics>=1.5.2",
]

[tool.uv.sources]
efficientvit = { git = "https://github.com/mit-han-lab/efficientvit.git" }
